django/admin/requirements.txt
asgiref==3.5.2
attrs==21.4.0
black==22.6.0
click==8.1.3
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pika==1.2.1
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1